Implement logging and code quality optimizations
- Configure NLog to skip EF Core SQL statement logging
- Modify cleanup service to output messages to console only
- Modify statistics service to output messages to console only
- Add slow query detection to QueryOptimizationExtensions
- Replace double with decimal for precision (FailureRatio, calculation fields)
- Add index monitoring and optimization suggestions

+++ ZakYip.Sorting.RuleEngine.Infrastructure/Resilience/ResiliencePolicyFactory.cs
@@ -67,7 +67,7 @@ public static class ResiliencePolicyFactory
     /// <returns>弹性管道</returns>
     public static ResiliencePipeline CreateApiCircuitBreakerPolicy(
         ILogger logger,
-        double failureThreshold = 0.5,
+        decimal failureThreshold = 0.5m,
         int samplingDuration = 60,
         int minimumThroughput = 10,
         int durationOfBreak = 60)
@@ -75,7 +75,7 @@ public static class ResiliencePolicyFactory
         return new ResiliencePipelineBuilder()
             .AddCircuitBreaker(new CircuitBreakerStrategyOptions
             {
-                FailureRatio = failureThreshold,
+                FailureRatio = (double)failureThreshold,
                 SamplingDuration = TimeSpan.FromSeconds(samplingDuration),
                 MinimumThroughput = minimumThroughput,
                 BreakDuration = TimeSpan.FromSeconds(durationOfBreak),
